Add support for INACTIVE action resume
This adds another method to the APT blueprint for doing a resume
operation. This is to be called to signal that an INACTIVE action
status may be resumed. This includes the case where the action is
INACTIVE because it requires additional consent which should be
signaled to the user by using the `ConsentRequired` value on the
`code` of the action status details returned and setting the
`required_scope` value to the scope the user should consent to
recover.

To help with setting proper details on `INACTIVE` state and on
`FAILED` states helper model classes have been added for encoding the
correct properties on FAILED and INACTIVE action statuses. These are
intended to be extensible by the user, but should include the `code`
and `description` fields. Further extending these for particular
cases (e.g. the ConsentRequired INACTIVE cause) could/should be
considered.