Change the behavior of caching for Auth/Groups calls
- The scope of the caches is now broadened beyond a single HTTP
transaction. Previously, the key to the maps used for caching was the
`AuthState` object which only has lifetime within a single HTTP
transaction. Thus, cache lifetime was also only a single HTTP
transaction which severely limited its usefulness as the idea of
caching is to avoid Auth/Groups calls across multiple requests.

- Caches are now based on `TTLCache` and `LRUCache` from cachetools (an added
dependency). This handles timeouts on cache entries for us. At
present, the timeout values for the various caches are set as follows:

  - Token introspect: 30 seconds
  - Dependent Token lookup: no timeout (`LRUCache` used) cache value
     considered valid as long as the dependent_tokens_cache_id from
     introspection is the same.
  - Group Membership lookup: 5 minutes

- It is also worth noting how these caches are keyed:

  - Token Introspect: the bearer token
  - Dependent Token: the value of `dependent_tokens_cache_id` returned
  from token introspection
  - Group Membership: The access token used for accessing the groups
  service (which is returned as a dependent token for the groups API scope)

- As we are now using cachetools, all usage of dogpile
cache has been removed. We've never needed the increased
functionality (with the increased complexity) it brings in.

  - If we were ever to bring in a shared caching solution like
  memcached or some such, it might be worth bringing back, but should
  evaluate it at that time.

The documentation related to caching in `caching.rst` has been changed
to describe these policies and removing all of the information
relating to configuring dogpile.