Add optimization for principal check without groups
Add the optimization wherein if we are doing an authorization check
but the list we're checking against doesn't include any groups we
won't look for the user's groups. This can avoid a round-trip to the
groups service and, potentially, even a dependent token exchange if
the only dependent service for a particular operation would have been
groups.

Also, this adds to the dev dependencies types required for linting to succeed.