Fix for missing or improperly cached refresh tokens
The previous implementation was checking for presence of the
`refresh_token` key in the dependent token response object. It turns
out that doing an `in` check always returns true, so this changes to
checking for the value of `None` rather than doing an `in` check. In the
previous mode, we could attempt to create a `RefreshTokenAuthorizer`
with a `None` `refresh_token` which will always fail. This, in turn,
would not show up until an AP attempted to use the authorizer via an
SDK client, like TransferClient, which could make it hard to trace
back to the root cause.

We also change the dependent token cache to a `TTLCache` with timeout
of 47 hours instead of a basic `LRUCache` so that we don't keep tokens
around longer than the expected expiration any way. Thus, the
expectation is that in the common case, prior to any tokens expiring
and attempting a refresh, we will, instead, no retain the tokens which
are due to expire and cause a new dependent grant. This is somewhat
non-optimal assuming dependent grants are more expensive, particularly
for Auth, than refresh operations, but it also makes handling missing
refresh tokens in some way automatic since we don't use them as often.

To that last, the source for a lot of this is that the dependent
grants do not seem to be returning refresh tokens at all in some
cases. That led to the whole path of building this additional
robustness that we didn't expect to need for missing refresh
tokens. This robustness here is the right thing, but why refresh
tokens are not coming in some cases is still an unsolved issue.