Updated get_authorizer_for_scope
This changes the behavior / priority of the type of authorizer to be
returned. If an access token is available, and the user hasn't
indicated that the access token needs validity longer than a requested
interval, we always return the AccessTokenAuthorizer. This avoids the
overhead of tracking auth clients and so on as part of the state of
the RefreshTokenAuthorizer.

If a refresh token is present, it will be
used to create a RefreshTokenAuthorizer which will, in turn, refresh
the access token which we by now know is likely to expire within a
user requested time window.

Finally, if we cannot create a RefreshTokenAuthorizer, we will recurse
to try the process again, only this time we will instruct to bypass
any dependent token caching (where the dependent scope access token
has been waiting and is likely to expire). This may update the cache
if a new dependent token is generated.