Replace GroupsClient with globus_sdk.GroupsClient
The custom GroupsClient object is removed, usages are updated to use
`globus_sdk.GroupsClient` as appropriate, and the scope string for the
view_my_groups_and_memberships scope is pulled from the SDK as well.

To support this, update the minimum globus-sdk version to `3.9.0` and
run `poetry lock --no-update`. v3.9.0 introduced the ArrayResponse
handling for `get_my_groups` and is therefore needed in order for the
response from that call to be iterable.

- No change is made to the way that testing is done (e.g. to use
  `globus_sdk._testing`)
- Some flake8-identified unused names (imports and vars) were removed
- No changelog entry is included, as this should be purely internal
  (the GroupsClient class is not exported in the top-level
  `__init__.py`)