Remove deprecation warning
This change makes no comment on the desirability of this module, but
it was deprecated and then *not* removed, and is still relied upon for
examples.

Without clarity on its true deprecation status, the sensible near-term
course is to remove this warning as "spurious", even if it reflects
some real desire to remove the module.