Establish dependencies among the tox environments
This allows tox to run the environments in parallel.
This is a big performance gain:

Serial tox runs (with all environments already built) take ~20 seconds.
Parallel tox runs now take only ~5 seconds to complete.

This requires a change to the `whattimeisitnow` example app.
Previously it was using an on-disk file, which led to conflicts.
Now, it's using an in-memory dictionary as its "database".