Replace 'TokenChecker' with 'AuthStateBuilder'
This change replaces 'TokenChecker' wherever it appears.

This fixes some bad naming, but it also moves the confidential client
construction *out of* the state builder, so that it is more visible
and can examined as a location which needs attention in order for us
to tune and control network behaviors.