Remove the 'testing' module
This is a *large* surface area exposed via a very small number of
low-utility test helpers. The contracts here cannot be maintained as
the main source itself changes, and therefore each change to the
source requires an attendant change to the testing helpers and
documentation of the change in both locations.

Move the only "surviving" test in `test_mocks` to a new `test_meta`
test module.