Remove 'aud' field checking and fields (#171)
Removing audience checking is safe, in that we remain fully spec
compliant. The field is optional for servers and optional for clients
to validate even when it's present.

The source for the expected audience value is typically the same as
that of the client credentials, and therefore there is no additional
safety (e.g., against credential confusion bugs) being added by
requiring this field. It only serves to add surface area and
therefore complexity.

This is a breaking change, in that the interfaces for the library are
changing to remove a field.