Refactor groups caching for simplicity and perf
This order of evaluation is cleaner/simpler to reason about and has
improved performance in specific cases, when the cache is warm and a
dependent token is absent (e.g., the group cache was populated at the
end of the lifetime for a cached dependent token).

This changeset refrains from changing the established behavior of
converting groups API errors to the empty group set, but it does make
this behavior more visible.