Remove flask 'api_helpers' and examples
This module is never used by consumers of the package, based on an
examination of known users. It was slated for removal in a much
earlier version 0.12.0, but the deprecation warning which was put in
place for this was incorrectly applied, so that all usage of the
library emitted the deprecation warning.
The warning was removed in 0.13.0 in order to fix this, and never
re-applied for simplicity.

After removing the module and tracing references to its functions, we
find that two example applications (unmaintained) in the `examples/`
directory were using these. Remove them.

Update the examples pages in the docs to accord with this removal.
Also, fix any description of the custom Flask Blueprint as
"Flask Decorators". This is a narrow characteristic of the Blueprint
object, and not an appropriate name for the feature.