Introduce a ClientFactory pattern encoding params
A ClientFactory is a newly-defined interface specific to AP Tools
which can construct ConfidentialAppAuthClient and GroupsClient
instances with pre-baked initialization parameters. In our case, the
default implementation is customized to provide transport-layer
parameters tuning the retry and timeout behaviors.

A user can implement the same interface and pass an alternative
ClientFactory to the blueprint config, allowing them freedom
to more tightly control the construction of Auth and Groups clients
within AP Tools.

Minimal changes are made to add `client_factory` as an init-time
argument for AuthStateBuilder and AuthState, and `client_factory` is
added as an attribute of the blueprint config (ActionProviderConfig).
The blueprint will pass its configured factory to AuthStateBuilder,
and the AuthStateBuilder will pass its configured factory to
AuthState.