Eliminate use of the `^` syntax for dependencies
`^` is a special symbol for Poetry, not part of the packaging
specifications. As a result, it is always being translated when we do
package builds, and may follow unintuitive rules.

Replace it with standardized package specifiers, which make our
statically declared metadata match the rendered PKG-INFO and wheel
metadata.

A build confirms that the current specifiers are very slightly
different. For example, globus-sdk was `>=3.9.0,<4.0.0`. This is
semantically the same as what we have written for almost all cases. In
theory, there could be some corner cases around values like `4.0.0a1`
and other such values, but none are obvious or likely.

`poetry lock --no-update` makes no change to the lockfile, confirming
that the values have not moved significantly.