Rewrite the `setting_up_auth` document
Every line of the doc has been evaluated and for the most part
replaced. The doc is more prescriptive and based around use of
`globus_sdk` in order to provide a good and ergonomic developer
experience.

Concepts are explained *minimally* but the focus is on getting the
client created and its scope created. Because this document does not
cover actually configuring an AP using the library, no mention is made
whatsoever of the `globus_auth_scope` field. However, we do note that
there is exactly one scope per AP. (Although not technically a strict
requirement, it is a practical fact, and mixing scopes from multiple
resource servers would cause failures.)

In the course of the revised guide, we walk the user through
constructing a `manage-ap.py` script which can act as their admin
interface for their AP. This will let them manage their scope as their
application evolves.

Following a best practice from the broader Python community, this doc
introduces some tabbed layout (using `sphinx-design`) in order to
provide instructions using the `py` runner for environment creation on
Windows. (See the `pip` documentation for other examples of this
practice.) In order to add `sphinx-design`, I have updated the nested
'docs' project and used `uvx --from 'poetry==1.8.4' poetry ...` to get
semantics which exactly match the version which was used when those
files were created and minimize the changes introduced here.