Revise tests to more consistently use `responses`
Rather than using a mocked `AuthState` object, our preference is to
mock HTTP with `responses` and use a real instance of the object. This
leads to more effective testing.

A discrepancy was revealed in one of the test fixtures in which the
effective identity ID was not included in the list of all identities
in the mocked token introspect data.

`RegisteredResponse.replace()` is used to put default success mocks in
place and replace them as necessary for specific tests.